PackageBuilder split out to netstandard
