+++ MSBuild.Sdk.SqlProj.slnx
@@ -13,6 +13,7 @@
   </Folder>
   <Folder Name="/src/">
     <Project Path="src/DacpacTool/DacpacTool.csproj" />
+    <Project Path="src/DacpacToolLibNetstandard/DacpacToolLibNetstandard.csproj" />
     <Project Path="src/MSBuild.Sdk.SqlProj/MSBuild.Sdk.SqlProj.csproj" />
   </Folder>
   <Folder Name="/test/">
