+++ src/DacpacTool/DacpacTool.csproj
@@ -15,6 +15,7 @@
     <PackageReference Include="Microsoft.SqlServer.DacFx" Version="170.3.93" />
     <PackageReference Include="NuGet.Versioning" Version="7.3.0" />
     <PackageReference Include="DotMake.CommandLine" Version="3.1.0" />
+    <ProjectReference Include="..\DacpacToolLibNetstandard\DacpacToolLibNetstandard.csproj" />
   </ItemGroup>
 
   <!-- References to Microsoft.SqlTools.ManagedBatchParser -->
