--- src/DacpacTool/IConsole.cs
@@ -1,8 +1,0 @@
-namespace MSBuild.Sdk.SqlProj.DacpacTool
-{
-    public interface IConsole
-    {
-        void WriteLine(string value);
-        string ReadLine();
-    }
-}
