+++ src/DacpacTool/ScriptParser.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using Microsoft.SqlServer.Dac.Model;
 using Microsoft.SqlTools.ServiceLayer.BatchParser;
+using MSBuild.Sdk.SqlProj.DacpacToolLibNetstandard;
 
 namespace MSBuild.Sdk.SqlProj.DacpacTool
 {
