+++ src/DacpacToolLibNetstandard/IConsole.cs
@@ -1,0 +1,7 @@
+namespace MSBuild.Sdk.SqlProj.DacpacToolLibNetstandard;
+
+public interface IConsole
+{
+    void WriteLine(string value);
+    string ReadLine();
+}
