+++ test/DacpacTool.Tests/PackageAnalyzerTests.cs
@@ -7,6 +7,7 @@ using System.Runtime.CompilerServices;
 using Microsoft.SqlServer.Dac.Model;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Shouldly;
+using MSBuild.Sdk.SqlProj.DacpacToolLibNetstandard;
 
 namespace MSBuild.Sdk.SqlProj.DacpacTool.Tests
 {
